+++ src/components/Header.jsx
@@ -1,5 +1,7 @@
+import Notifications from "./Notifications";
 import Scoreboard from "./Scoreboard";
 import GithubIcon from "../assets/icons/github-mark/github-mark.svg?react";
+import PlayfulCatIcon from "../assets/icons/miscellaneous/undraw_playful_cat.svg?react";
 import "../styles/header.css";
 
 export default function Header({ currentScore, bestScore }) {
@@ -9,16 +11,18 @@ export default function Header({ currentScore, bestScore }) {
         <ul className="nav-left">
           <li>
             <a href="#">
-              <img src="#"></img>
+              <PlayfulCatIcon className="icon logo" alt="#" />
               <h1>Memory Card</h1>
             </a>
           </li>
         </ul>
+        <Notifications currentScore={currentScore} />
         <Scoreboard currentScore={currentScore} bestScore={bestScore} />
+
         <ul className="nav-right">
           <li>
             <a href="#" target="_blank">
-              <GithubIcon className="icon github" />
+              <GithubIcon className="icon github" alt="#" />
             </a>
           </li>
         </ul>