+++ src/components/Main.jsx
@@ -3,16 +3,24 @@ import Cards from "./Cards";
 import "../styles/main.css";
 
 export default function Main({
+  gameWinRef,
+  newGame,
+  gameOver,
   currentScore,
   bestScore,
+  setGameOver,
   setCurrentScore,
   setBestScore,
 }) {
   return (
     <main>
       <Cards
+        gameWinRef={gameWinRef}
+        newGame={newGame}
+        gameOver={gameOver}
         currentScore={currentScore}
         bestScore={bestScore}
+        setGameOver={setGameOver}
         setCurrentScore={setCurrentScore}
         setBestScore={setBestScore}
       />